wood: separate ext_l feasibility from LIMIT_MIN_JOINT_LENGTH, add fold executables
- wood_face_to_face: independent ext_sq and min_length checks
- main_wood_fold_1: miter contour plates with joint volume extension
- mesh, line, vector, polyline: various geometry updates
- wood_joint, wood_merge, wood_main: joint library and merge improvements

+++ src/yaml/yaml.cpp
@@ -245,7 +245,7 @@ namespace TINY_YAML {
 						break;
 					}
 					/*Make the current node the new parent*/
-					parentsStack.push(Triple<std::shared_ptr<Node>, unsigned int, bool>(pnode, dashPos, true));
+					parentsStack.push(Triple<std::shared_ptr<Node>, unsigned int, bool>(pnode, static_cast<unsigned int>(dashPos), true));
 					dashPos = std::string::npos;	
 				}
 				else { /* A list of elements inside the current parent pnode */
@@ -275,7 +275,7 @@ namespace TINY_YAML {
 				else if (parentsStack.size() != 0 && !parentsStack.top().first->append(pnode)) { // If it is not at root level and it is failed to attach the current node to the current parent
 					faulty = true; break;
 				}
-				parentsStack.push(Triple<std::shared_ptr<Node>, unsigned int, bool>(pnode, firstCharPos, false));
+				parentsStack.push(Triple<std::shared_ptr<Node>, unsigned int, bool>(pnode, static_cast<unsigned int>(firstCharPos), false));
 				continue;
 			}
